swift: wait for old objects to be deleted after move calls
This is necessary due to eventual consistency and to decrease the chance
of it being deleted again in calls to (*driver).Delete().

This is specially concerning for large objects, because moved objects
still share the same segments if (*driver).Delete is called on them and
swift still lists the segments for the moved object we'll be deleting
segments needed by the move destination object, making it look
corrupted.